# projects/agent-kanban/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Agent Kanban API
  description: |
    多项目 AI Agent 统一看板的 REST API。

    核心资源链：Workspace → Project → Task → Assignment → Agent

    **契约说明**：
    - Task 的创建/更新/关闭以 GitHub Issues 为准，本 API 提供只读聚合查询。
    - Assignment 的创建和管理通过本 API。
    - 此文件是 **canonical** API 契约，Go 和 Node.js 两个后端都基于此实现。
    - 所有时间字段为 ISO 8601 格式 (e.g. "2026-08-03T07:22:00Z")。
  version: "1.1.0"
  contact:
    name: Agent Kanban Team
servers:
  - url: http://localhost:8080/api/v1
    description: Go 后端 (本地开发)
  - url: http://localhost:4001/api/v1
    description: Node.js 后端 (本地开发)
  - url: http://localhost:4002/api/v1
    description: Next.js 嵌入式后端 (本地开发)
security:
  - ApiKeyAuth: []
  - {}
...
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      summary: 获取所有工作空间
      operationId: listWorkspaces
      responses:
        "200":
          description: 工作空间列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
              examples:
                default:
                  value:
                    - id: "ws-001"
                      name: "产品团队"
                      description: "负责所有产品线开发"
                      userId: null
                      createdAt: "2026-08-01T00:00:00Z"
                      updatedAt: "2026-08-01T00:00:00Z"
    post:
      tags: [Workspaces]
      summary: 创建工作空间
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceInput"
            examples:
              default:
                value:
                  name: "产品团队"
                  description: "负责所有产品线开发"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
              examples:
                default:
                  value:
                    id: "ws-001"
                    name: "产品团队"
                    description: "负责所有产品线开发"
                    userId: null
                    createdAt: "2026-08-03T07:22:00Z"
                    updatedAt: "2026-08-03T07:22:00Z"
        "400":
          $ref: "#/components/responses/ValidationError"
  /workspaces/{workspaceId}:
    get:
      tags: [Workspaces]
      summary: 获取工作空间详情
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: 工作空间详情（含项目和 Agent 列表）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Workspaces]
      summary: 删除工作空间（级联删除项目和 Agent）
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "204":
          description: 删除成功
        "404":
          $ref: "#/components/responses/NotFound"
  /workspaces/{workspaceId}/stats:
    get:
      tags: [Stats]
      summary: 获取工作空间统计概览
      operationId: getWorkspaceStats
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: 统计数据
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceStats"
              examples:
                default:
                  value:
                    totalTasks: 42
                    backlogTasks: 10
                    inProgressTasks: 8
                    inReviewTasks: 5
                    doneTasks: 19
                    totalAgents: 3
                    activeAgents: 2
                    totalProjects: 2
                    tasksByProject:
                      "前端项目": 25
                      "后端项目": 17
                    tasksByAgent:
                      "前端专家": 12
                      "后端专家": 8

  # ═══════════════════════════════════════════════
  # Projects
  # ═══════════════════════════════════════════════
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 可选。Zero-Auth 模式下不需要。生产环境启用时在 header 传 X-API-Key。
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            default:
              value:
                code: "NOT_FOUND"
                message: "Workspace not found"
    ValidationError:
      description: 请求参数验证失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            default:
              value:
                code: "VALIDATION_ERROR"
                message: "name is required"
                details:
                  field: "name"
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [NOT_FOUND, VALIDATION_ERROR, CONFLICT, UNAUTHORIZED, NO_MATCHING_AGENT, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: object
...
    Workspace:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        userId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateWorkspaceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
    WorkspaceDetail:
      allOf:
        - $ref: "#/components/schemas/Workspace"
        - type: object
          properties:
            projects:
              type: array
              items:
                $ref: "#/components/schemas/Project"
            agents:
              type: array
              items:
                $ref: "#/components/schemas/Agent"
    WorkspaceStats:
      type: object
      properties:
        totalTasks:
          type: integer
        backlogTasks:
          type: integer
        inProgressTasks:
          type: integer
        inReviewTasks:
          type: integer
        doneTasks:
          type: integer
        totalAgents:
          type: integer
        activeAgents:
          type: integer
        totalProjects:
          type: integer
        tasksByProject:
          type: object
          additionalProperties:
            type: integer
        tasksByAgent:
          type: object
          additionalProperties:
            type: integer
    Project:
      type: object
      required: [id, workspaceId, name, repoOwner, repoName, syncStatus, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        repoOwner:
          type: string
        repoName:
          type: string
        webhookSecret:
          type: string
          nullable: true
          description: 仅在创建时返回一次
        syncStatus:
          type: string
          enum: [idle, syncing, error]
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Agent:
      type: object
      required: [id, workspaceId, name, role, modelName, capabilityTags, status, maxConcurrency]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        role:
          type: string
        description:
          type: string
        modelName:
          type: string
        capabilityTags:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        githubBotAccount:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, paused, offline]
        maxConcurrency:
          type: integer
          default: 3
        createdAt:
          type: string
          format: date-time
...